PIM-3832: move product import classes in new component and bundle
