+++ src/Pim/Bundle/ConnectorBundle/Doctrine/ORM/Repository/JobConfigurationRepository.php
@@ -11,8 +11,6 @@ use Pim\Component\Connector\Repository\JobConfigurationRepositoryInterface;
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * TODO: move this in a Doctrine folder
  */
 class JobConfigurationRepository extends EntityRepository implements JobConfigurationRepositoryInterface
 {
