+++ src/Pim/Bundle/ConnectorBundle/Resources/config/processors.yml
@@ -1,8 +1,11 @@
 parameters:
     pim_base_connector.processor.denormalization.attribute_option.class:                  Pim\Component\Connector\Processor\Denormalization\AttributeOptionProcessor
+    pim_base_connector.processor.denormalization.product.class:                           Pim\Component\Connector\Processor\Denormalization\ProductProcessor
+    pim_base_connector.processor.denormalization.product_association.class:               Pim\Component\Connector\Processor\Denormalization\ProductAssociationProcessor
 
     pim_base_connector.denormalization.array_converter.flat.attribute_option.class:       Pim\Component\Connector\Processor\Denormalization\ArrayConverter\Flat\AttributeOptionToStandardConverter
     pim_base_connector.denormalization.array_converter.structured.attribute_option.class: Pim\Component\Connector\Processor\Denormalization\ArrayConverter\Structured\AttributeOptionToStandardConverter
+    pim_base_connector.denormalization.array_converter.flat.product.class:                Pim\Component\Connector\Processor\Denormalization\ArrayConverter\Flat\ProductToStandardConverter
 
 services:
     # Denormalization - Processors
@@ -22,6 +25,27 @@ services:
             - '@pim_catalog.updater.attribute_option'
             - %pim_catalog.entity.attribute_option.class%
 
+    pim_base_connector.processor.denormalization.product.flat:
+        class: %pim_base_connector.processor.denormalization.product.class%
+        arguments:
+            - '@pim_base_connector.denormalization.array_converter.flat.product'
+            - '@pim_catalog.repository.product'
+            - '@pim_catalog.builder.product'
+            - '@pim_catalog.updater.product'
+            - '@pim_validator'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+
+    pim_base_connector.processor.denormalization.product_association.flat:
+        class: %pim_base_connector.processor.denormalization.product_association.class%
+        arguments:
+            - '@pim_catalog.repository.product'
+            - '@pim_serializer'
+            - '@pim_validator'
+            - '@pim_transform.builder.field_name'
+            - %pim_catalog.entity.association.class%
+            - %pim_catalog.entity.product.class%
+            - 'csv'
+
     # Denormalization - ArrayConverters
     pim_base_connector.denormalization.array_converter.flat.attribute_option:
         class: %pim_base_connector.denormalization.array_converter.flat.attribute_option.class%
@@ -30,3 +54,11 @@ services:
 
     pim_base_connector.denormalization.array_converter.structured.attribute_option:
         class: %pim_base_connector.denormalization.array_converter.structured.attribute_option.class%
+
+    pim_base_connector.denormalization.array_converter.flat.product:
+        class: %pim_base_connector.denormalization.array_converter.flat.product.class%
+        arguments:
+            - '@pim_transform.builder.field_name'
+            - '@pim_catalog.repository.attribute'
+            - '@pim_catalog.repository.currency'
+            - '@pim_catalog.resolver.attribute_values'
