+++ src/Pim/Component/Connector/Processor/Denormalization/ProductAssociationProcessor.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Processor\Denormalization;
+namespace Pim\Component\Connector\Processor\Denormalization;
 
 use Akeneo\Bundle\StorageUtilsBundle\Repository\IdentifiableObjectRepositoryInterface;
+use Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\AbstractProcessor;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\TransformBundle\Builder\FieldNameBuilder;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
